+++ src/routes/(app)/+page.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-	import HeroSection from './_components/HeroSection.svelte';
-</script>
-
-<div class="flex grow flex-col items-center justify-center">
-	<HeroSection />
-</div>