+++ src/routes/(app)/submissions/+page.svelte
@@ -4,7 +4,6 @@
 	import Avatar from '$lib/components/Avatar.svelte';
 	import { Pagination } from 'bits-ui';
 	import { ChevronLeftIcon, ChevronRightIcon } from 'lucide-svelte';
-	import { _ } from 'svelte-i18n';
 	import type { PageData } from './$types';
 	import FilterBar from './_components/FilterBar.svelte';
 	import StatusIcon from './_components/StatusIcon.svelte';
@@ -32,27 +31,25 @@
 	};
 </script>
 
-<h1 class="text-2xl font-bold">{$_('submissions', { default: 'Submissions' })}</h1>
-
-<div class="w-full px-xl py-sm">
+<div class="w-full px-sm md:px-xl">
 	<FilterBar {events} {categories} />
 </div>
-<div class="grid w-full grid-cols-submissions gap-sm p-xl">
+<div class="grid w-full grid-cols-submissions gap-sm px-sm md:gap-xl md:px-xl">
 	{#each submissions as submission}
 		<div class="wrapper group grid overflow-hidden rounded-md">
-			<div class="h-full w-full overflow-hidden rounded-md">
+			<div class="aspect-square h-full w-full overflow-hidden rounded-md">
 				<a href={`/submissions/${submission.id}`}>
 					{#if submission?.media?.type === 'image'}
 						<img
 							src={`${submission?.media?.url}`}
 							alt={submission?.media?.alt}
-							class="w-full object-cover object-center"
+							class="h-full w-full object-cover object-center"
 						/>
 					{:else if submission?.media?.type === 'audio'}
 						<img
 							src={`/images/music-thumbnail.webp`}
 							alt="Audio"
-							class="w-full object-cover object-center"
+							class="h-full w-full object-cover object-center"
 						/>
 					{/if}
 				</a>
@@ -67,7 +64,7 @@
 				{/if}
 				<a
 					href={`/users/${submission.user?.username}`}
-					class="pointer-events-auto inline-flex h-form w-fit -translate-y-sm items-center gap-sm rounded-full p-xs pr-md opacity-0 transition-all duration-300 group-hover:translate-y-0 group-hover:bg-black/75 group-hover:opacity-100 group-hover:backdrop-blur"
+					class="pointer-events-auto inline-flex h-form w-fit translate-y-0 items-center gap-sm rounded-full p-xs pr-md transition-all duration-300 group-hover:bg-black/75 group-hover:backdrop-blur md:-translate-y-sm md:opacity-0 md:group-hover:translate-y-0 md:group-hover:opacity-100"
 				>
 					<Avatar
 						src={submission.user?.picture ?? ''}
@@ -77,7 +74,7 @@
 					<p>{submission.user?.username}</p>
 				</a>
 				<div
-					class="translate-y-sm rounded-md p-sm opacity-0 transition-all duration-300 group-hover:translate-y-0 group-hover:bg-black/75 group-hover:opacity-100 group-hover:backdrop-blur"
+					class="rounded-md bg-black/75 p-sm backdrop-blur transition-all duration-300 group-hover:backdrop-blur md:translate-y-sm md:bg-none md:opacity-0 md:backdrop-blur-none md:group-hover:translate-y-0 md:group-hover:bg-black/75 md:group-hover:opacity-100"
 				>
 					<h2 class="text-lg font-bold">{submission.title}</h2>
 					<p class="text-muted-foreground">{submission.category?.name}</p>