+++ src/routes/(app)/submissions/[id]/+page.svelte
@@ -27,7 +27,7 @@
 	);
 </script>
 
-<main class="flex w-full max-w-screen-lg flex-col gap-lg p-sm">
+<main class="flex w-full max-w-screen-lg flex-col gap-sm px-sm pt-sm md:gap-xl md:px-xl md:pt-xl">
 	<!-- Header -->
 	<div class="flex flex-wrap items-center justify-between gap-xs">
 		<h1 class="text-2xl font-bold">{submission?.title}</h1>
@@ -59,7 +59,7 @@
 					</Button>
 				</Popover.Trigger>
 				<Popover.Content
-					class="z-50 max-h-[320px] w-fit max-w-[320px] overflow-hidden overflow-y-auto rounded-lg bg-muted-background p-sm"
+					class="z-30 max-h-[320px] w-fit max-w-[320px] overflow-hidden overflow-y-auto rounded-lg bg-muted-background p-sm"
 					sideOffset={5}
 				>
 					{#if user == null}