+++ src/routes/+page.server.ts
@@ -1,5 +1,9 @@
+import { redirect } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
 
 export const load = (async ({ locals }) => {
-	return { user: locals.user };
+	if (locals.user) {
+		throw redirect(302, '/submissions');
+	}
+	return {};
 }) satisfies PageServerLoad;