+++ src/routes/+page.svelte
@@ -1,0 +1,20 @@
+<script lang="ts">
+	import Footer from './(app)/_components/Footer.svelte';
+	import HeroSection from './(app)/_components/HeroSection.svelte';
+	import Navbar from './(app)/_components/Navbar.svelte';
+</script>
+
+<svelte:head>
+	<title>Kreacon</title>
+</svelte:head>
+
+<div class="flex min-h-screen flex-col">
+	<Navbar />
+
+	<div class="flex min-h-full grow flex-col items-center">
+		<div class="flex grow flex-col items-center justify-center">
+			<HeroSection />
+		</div>
+	</div>
+	<Footer />
+</div>